feat: Improve add performances (#22)
* feat: improve add performances

Re use the entityRecord instead of fetching it in all sub functions

* fix: calling AddComponentsN moves the archetype

AddComponentsN were not meant to be used multiple time on the same
entity. But in the Arche benchmark it is required to add 10 components
(AddComponents8 and AddComponents3). This resulted to the entity having
only 3 components and being attache to the archetype of those 3
components. The 8 other components were lost in memory, in their
storage.